requirements.txt
absl-py==2.0.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.3
func-timeout==4.3.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
gym==0.26.2
gym-notices==0.0.8
idna==3.4
imageio==2.32.0
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
pandas==2.1.3
Pillow==10.0.1
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-video==1.1.11
scipy==1.11.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tifffile==2023.9.26
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1